Implement the Marker on maps
